+++ lib/src/routes_tree.dart
@@ -11,11 +11,15 @@ class RoutesTree {
   QRouterDelegate _rootDelegate;
 
   // Build the Route Tree.
-  List<_QRoute> _buildTree(List<QRoute> routes, String basePath, int key) {
+  List<_QRoute> _buildTree(List<QRouteBase> routes, String basePath) {
     final result = <_QRoute>[];
     if (routes == null || routes.isEmpty) return result;
 
-    for (var route in routes) {
+    for (var routebase in routes) {
+      var route = routebase is QRoute
+          ? routebase
+          : (routebase as QRouteBuilder).createRoute();
+
       var path = route.path;
       if (!path.startsWith('/')) {
         path = '/$path';
@@ -39,7 +43,7 @@ class RoutesTree {
       }
       final fullPath = basePath + (path.isEmpty ? '' : '/$path');
       final _route = _QRoute(
-        key: key + routes.indexOf(route),
+        key: _routesIndex.length + 1,
         name: route.name,
         isComponent: path.startsWith(':'),
         path: path,
@@ -49,27 +53,23 @@ class RoutesTree {
         redirectGuard: route.redirectGuard ?? (s) => null,
         fullPath: fullPath,
       );
-      _route.children
-          .addAll(_buildTree(route.children, fullPath, key + routes.length));
+      _route.children.addAll(_buildTree(route.children, fullPath));
       result.add(_route);
-      key += routes.indexOf(route);
-      if (route.name != null) {
-        _routesIndex[route.name] = fullPath;
-      }
-      QR.log('"${_route.name}" added with base $basePath');
+      _routesIndex[route.name ?? _route.fullPath] = fullPath;
+      QR.log('"${_route.name}" added with path ${_route.fullPath}');
     }
     return result;
   }
 
   QRouterDelegate setTree(
-      List<QRoute> routes, QRouterDelegate Function() delegate) {
+      List<QRouteBase> routes, QRouterDelegate Function() delegate) {
     if (_routes.isNotEmpty) {
       QR.log('Reset Tree');
       _routes.clear();
       _routesIndex.clear();
     }
 
-    _routes.addAll(_buildTree(routes, '', 1));
+    _routes.addAll(_buildTree(routes, ''));
     for (var route in _routes) {
       route.printTree(1);
     }
@@ -82,8 +82,8 @@ class RoutesTree {
     QR.log('matching for $path', isDebug: true);
 
     // Check if the same route
-    if (path == QR.currentRoute.fullPath && QR.currentRoute.match != null) {
-      return QR.currentRoute.match;
+    if (path == QR.currentRoute.fullPath && _cureentTree != null) {
+      return _cureentTree;
     }
 
     final match = _getMatch(path);
@@ -148,7 +148,6 @@ class RoutesTree {
     _cureentTree = newTree;
     QR.currentRoute.fullPath = path;
     QR.currentRoute.params = match.getParames();
-    QR.currentRoute.match = newTree;
     QR.history.add(path);
     return newTree;
   }
@@ -234,7 +233,7 @@ class _QRoute {
   final Function onInit;
   final RedirectGuard redirectGuard;
   final Function onDispose;
-  final QRouteBuilder page;
+  final QRoutePage page;
   final bool isComponent;
   final List<_QRoute> children = [];
 
@@ -255,6 +254,8 @@ class _QRoute {
         name: name ?? this.name,
         path: path ?? this.path,
         page: page,
+        onInit: onInit,
+        onDispose: onDispose,
         redirectGuard: redirectGuard,
         isComponent: isComponent);
     result.children.addAll(children);